+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * **Feature:** Added `std:process:spawn` and `std:process:kill_wait`.
+* **Feature:** Added `std:process:wait` for waiting it to exit.
 * **Feature**: Calling into EvalContext::eval\*() can now be done
 recursively.
 * **Feature:**: Added simple UDP networking via `std:net:udp:new`,